make seperate file for decorator class
