+++ capitalize_decorator.rb
@@ -1,7 +1,7 @@
-require './nameable'
+require './decorator'
 
 class CapitalizeDecorator < Decorator
   def correct_name
-    @nameable.correct_name.upcase
+    @nameable.correct_name.capitalize
   end
 end
