+++ main.rb
@@ -3,7 +3,7 @@ require './trimmer_decorator'
 require './capitalize_decorator'
 
 person = Person.new(22, 'maximilianus')
-person.correct_name
+p person.correct_name
 capitalized_person = CapitalizeDecorator.new(person)
 p capitalized_person.correct_name
 capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
