+++ nameable.rb
@@ -4,15 +4,3 @@ class Nameable
   end
 end
 
-class Decorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
